refactor(tds-1210): change package for daikon kafka mongo
