+++ daikon-kafka-mongo/src/main/java/org/talend/daikon/kafka/mongo/DaikonMongoConfiguration.java
@@ -1,10 +1,10 @@
-package org.talend.daikon.mongo;
+package org.talend.daikon.kafka.mongo;
 
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
-@ComponentScan(basePackages = { "org.talend.daikon.mongo" })
+@ComponentScan(basePackages = { "org.talend.daikon.kafka.mongo" })
 @Configuration
 @EnableAutoConfiguration
 public class DaikonMongoConfiguration {
