+++ daikon-kafka-mongo/src/main/java/org/talend/daikon/kafka/mongo/repo/MongoPendingRecordRepository.java
@@ -1,9 +1,9 @@
-package org.talend.daikon.mongo.repo;
+package org.talend.daikon.kafka.mongo.repo;
 
 import java.util.List;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
-import org.talend.daikon.mongo.repo.model.PersistentPendingRecord;
+import org.talend.daikon.kafka.mongo.repo.model.PersistentPendingRecord;
 import org.talend.daikon.repo.RecordPriority;
 
 public interface MongoPendingRecordRepository extends MongoRepository<PersistentPendingRecord, String> {
