+++ daikon-kafka-mongo/src/test/java/org/talend/daikon/kafka/mongo/util/TestRepositoryAbstract.java
@@ -1,8 +1,5 @@
-package org.talend.daikon.mongo.util;
+package org.talend.daikon.kafka.mongo.util;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
 import org.junit.After;
@@ -10,18 +7,17 @@ import org.junit.Rule;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.*;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.data.mongodb.config.AbstractMongoConfiguration;
 import org.springframework.data.mongodb.core.MongoTemplate;
-import org.springframework.data.mongodb.core.convert.CustomConversions;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
-import org.springframework.stereotype.Repository;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.talend.daikon.mongo.util.MongoCustomConverter;
-import org.talend.daikon.mongo.util.RepoConfiguration;
 
 import com.github.fakemongo.Fongo;
 import com.lordofthejars.nosqlunit.mongodb.MongoDbRule;
@@ -64,29 +60,17 @@ public abstract class TestRepositoryAbstract {
     }
 
     @Configuration
-    @EnableMongoRepositories(basePackages = "org.talend.daikon.mongo")
-    @ComponentScan(basePackages = "org.talend.daikon.mongo", useDefaultFilters = false, includeFilters = @ComponentScan.Filter(value = {
-            Repository.class, MongoCustomConverter.class }, type = FilterType.ANNOTATION))
+    @EnableMongoRepositories(basePackages = "org.talend.daikon.kafka.mongo")
+    @ComponentScan(basePackages = "org.talend.daikon.kafka.mongo")
     @Import(RepoConfiguration.class)
     static public class MongoTestConfiguration extends AbstractMongoConfiguration {
 
         private static String MONGO_DB = "daikonkafka";
 
-        @Autowired
-        private ApplicationContext applicationContext;
-
         @Override
         @Bean
         public Mongo mongo() throws Exception {
             return new Fongo(MONGO_DB).getMongo();
-        }
-
-        @Override
-        public CustomConversions customConversions() {
-            final List<Object> converters = new ArrayList<>();
-            Map<String, Object> beansWithAnnotation = applicationContext.getBeansWithAnnotation(MongoCustomConverter.class);
-            beansWithAnnotation.values().forEach(converters::add);
-            return new CustomConversions(converters);
         }
 
         @Override
